+++ scripts/auth/createNewUser.js
