+++ scripts/auth/getActiveUser.js
