+++ scripts/auth/loginInit.js
